Refactor PageNav and add LD flag
