+++ src/components/PageNav/PageNav.tsx
@@ -1,18 +1,34 @@
 import React from 'react'
 import { SideNav } from '@trussworks/react-uswds'
-
+import { useFlags } from 'launchdarkly-react-client-sdk'
 import styles from './PageNav.module.scss'
 
 import NavLink from 'components/util/NavLink/NavLink'
 
-type PropTypes = {
-  navItems: {
-    path: string
-    label: React.ReactNode
-  }[]
-}
+const PageNav = () => {
+  const flags = useFlags()
 
-const PageNav = ({ navItems }: PropTypes) => {
+  const navItems = [
+    { path: '/', label: 'My Space' },
+    {
+      path: '/sites-and-applications',
+      label: <>All sites &amp; applications</>,
+    },
+    { path: '/ussf-documentation', label: 'USSF documentation' },
+  ]
+
+  if (flags.guardianDirectory) {
+    // TODO: we can remove this if and put the navItem in the correct slot above once fully released
+    navItems.splice(1, 0, {
+      path: '/guardian-directory',
+      label: 'Guardian Directory',
+    })
+  }
+
+  if (flags.landingPageIndex) {
+    navItems.push({ path: '/landing', label: 'Landing Pages' })
+  }
+
   const items = navItems.map((i) => (
     <NavLink
       href={i.path}
