+++ src/layout/DefaultLayout/DefaultLayout.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react'
 import { GovBanner, GridContainer, Grid } from '@trussworks/react-uswds'
-import { useFlags } from 'launchdarkly-react-client-sdk'
 import styles from './DefaultLayout.module.scss'
 import Header from 'components/Header/Header'
 import PersonalData from 'components/PersonalData/PersonalData'
@@ -24,26 +23,10 @@ const DefaultLayout = ({
   rightSidebar?: JSX.Element
   children: React.ReactNode
 }) => {
-  const flags = useFlags()
   const { loading, portalUser } = useUser()
   const { initializeMySpace } = useMySpaceContext()
   const { searchQuery, setSearchQuery } = useSearchContext()
   const [displayName, setDisplayName] = useState<string>('')
-  const navItems = [
-    { path: '/', label: 'My Space' },
-    {
-      path: '/sites-and-applications',
-      label: <>All sites &amp; applications</>,
-    },
-    { path: '/ussf-documentation', label: 'USSF documentation' },
-  ]
-  if (flags.guardianDirectory) {
-    // TODO: we can remove this if and put the navItem in the correct slot above once fully released
-    navItems.splice(1, 0, {
-      path: '/guardian-directory',
-      label: 'Guardian Directory',
-    })
-  }
 
   useEffect(() => {
     if (portalUser) {
@@ -78,7 +61,7 @@ const DefaultLayout = ({
             <Grid row gap>
               {/* LEFT SIDEBAR */}
               <Grid tablet={{ col: displayFeedbackCard ? 3 : 2 }}>
-                <PageNav navItems={navItems} />
+                <PageNav />
                 {displayFeedbackCard && <FeedbackCard />}
               </Grid>
               <Grid tablet={{ col: true }}>
